+++ ostap/parallel/parallel_fill.py
@@ -13,7 +13,9 @@ __version__ = "$Revision$"
 __author__  = "Vanya BELYAEV Ivan.Belyaev@itep.ru"
 __date__    = "2011-06-07"
 __all__     = (
-    'pprocess'  , ## paralell processing 
+    'parallel_make_dataset' , ## efficient function for parallel fill of dataset 
+    'parallel_fill_dataset' , ## flexible  function for parallel fill of dataset 
+    'parallel_fill'         , ## flexible  function for parallel fill of dataset 
     ) 
 # =============================================================================
 from   ostap.parallel.parallel import Task, WorkManager
@@ -25,33 +27,33 @@ from ostap.logger.logger import getLogger
 if '__main__' ==  __name__ : logger = getLogger ( 'ostap.parallel.fill' )
 else                       : logger = getLogger ( __name__     )
 # =============================================================================
-## The simple task object for more efficient fill of RooDataSet from TChain 
+## The simplest task object for efficient fill of RooDataSet from TChain 
 #  @see GaudiMP.Parallel
 #  @see GaudiMP.Parallel.Task
-#  @see Ostap.SelectorWithVars
 #  For 12-core machine, clear speed-up factor of about 8 is achieved 
 #  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
 #  @date   2014-09-23 
-class  FillTask(Task) :
-    """The single task object for more efficient fill of RooDataSet from TChain 
+class  MakeDSTask(Task) :
+    """The simplest task object for efficient fill of RooDataSet from TChain 
     - for 12-core machine, clear speed-up factor of about 8 is achieved 
     """
     ## 
-    def __init__ ( self               ,
-                   variables          ,
-                   selection          ,
-                   roo_cuts  = ''     , 
-                   trivial   = False  , 
-                   use_frame = 100000 ) :
+    def __init__ ( self           ,
+                   variables      ,
+                   selection      ,
+                   roo_cuts  = '' ,
+                   name      = '' ,
+                   title     = '' ) :
         
-        self.variables = variables 
-        self.selection = selection
-        self.roo_cuts  = roo_cuts 
-        self.trivial   = trivial  
-        self.use_frame = use_frame 
-        self.__output  = ()  
+        self.variables  = variables 
+        self.selection  = selection
+        self.roo_cuts   = roo_cuts
+        self.name       = name
+        self.title      = title         
+        self.the_output = ()  
 
-    def initialize_local   ( self ) : self.__output = () 
+    ## local initialization 
+    def initialize_local   ( self ) : self.the_output = () 
 
     ## the actual processing 
     def process ( self , jobid , item ) :
@@ -62,7 +64,148 @@ class  FillTask(Task) :
             import ostap.core.pyrouts            
             import ostap.trees.trees
             import ostap.fitting.roofit 
+            import ostap.fitting.pyselectors
+        
+        ## reconstruct chain from the item 
+        chain    = item.chain        
+        return chain.make_dataset ( variables = self.variables ,
+                                    selection = self.selection ,
+                                    roo_cuts  = self.roo_cuts  ,
+                                    name      = self.name      ,
+                                    title     = self.title     ,                                             
+                                    silent    = True           ) 
+    
+    ## merge results/datasets 
+    def merge_results ( self , result , jobid = -1 ) :
+        """Merge results/datasets
+        """
+        import ostap.fitting.dataset
+        if result :
+            ds , stat = result
+            if not self.the_output or not self.the_output[0] :
+                self.the_output = ds , stat  
+            else :
+                ds_ , stat_ = self.the_output
+                ds_.append ( ds )
+                stat_.total      += stat.total     ## total 
+                stat_.processed  += stat.processed ## procesed 
+                stat_.skipped    += stat.skipped   ## skipped                
+                self.the_output   = ds_ , stat_
+                ds.erase () 
+            del result            
+            logger.debug ( 'Merging: %d entries ' % len( self.the_output[0] ) )
+        else :
+            logger.error ( "No valid results for merging" )
+            
+    ## get the results 
+    def results ( self ) :
+        return self.the_output
+    
+# =================================================================================
+## The simple task object for more efficient fill of RooDataSet from TChain 
+#  @see GaudiMP.Parallel
+#  @see GaudiMP.Parallel.Task
+#  @see Ostap.SelectorWithVars
+#  For 12-core machine, clear speed-up factor of about 8 is achieved 
+#  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+#  @date   2014-09-23 
+class  FillDSTask(MakeDSTask) :
+    """The single task object for more efficient fill of RooDataSet from TChain 
+    - for 12-core machine, clear speed-up factor of about 8 is achieved 
+    """
+    ## 
+    def __init__ ( self               ,
+                   variables          ,
+                   selection          ,
+                   cuts      = None   ,
+                   roo_cuts  = ''     ,
+                   name      = ''     ,
+                   title     = ''     , 
+                   shortcut  = True   , 
+                   use_frame = 100000 ) :
+        
+        MakeDSTask.__init__ ( self                  ,
+                             variables = variables , 
+                             selection = selection ,
+                             roo_cuts  = roo_cuts  ,
+                             name      = name      ,
+                             title     = title     )
+        
+        self.cuts      = cuts     
+        self.use_frame = use_frame 
+        self.shortcut  = shortcut
+        
+    ## the actual processing 
+    def process ( self , jobid , item ) :
 
+        import ROOT
+        from ostap.logger.logger import logWarning
+        with logWarning() :
+            import ostap.core.pyrouts            
+            import ostap.trees.trees
+            import ostap.fitting.roofit 
+            from   ostap.fitting.pyselectors import SelectorWithVars
+            
+        ## reconstruct chain from the item 
+        chain    = item.chain
+                
+        ## use selector  
+        selector = SelectorWithVars ( variables = self.variables ,
+                                      selection = self.selection ,
+                                      cuts      = self.cuts      ,
+                                      roo_cuts  = self.roo_cuts  ,
+                                      name      = self.name      ,
+                                      fuillname = self.title     , 
+                                      silence   = True           )
+        
+        return chain.fill_dataset2 ( selector  ,
+                                     silent    = True            , 
+                                     shortcut  = self.shortcut   ,
+                                     use_frame = self.use_frame  )
+        
+# =================================================================================
+## The simple task object for more efficient fill of RooDataSet from TChain 
+#  @see GaudiMP.Parallel
+#  @see GaudiMP.Parallel.Task
+#  @see Ostap.SelectorWithVars
+#  For 12-core machine, clear speed-up factor of about 8 is achieved 
+#  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+#  @date   2014-09-23 
+class  FillTask(MakeDSTask) :
+    """The single task object for more efficient fill of RooDataSet from TChain 
+    - for 12-core machine, clear speed-up factor of about 8 is achieved 
+    """
+    ## 
+    def __init__ ( self               ,
+                   variables          ,
+                   selection          ,
+                   cuts      = None   ,
+                   roo_cuts  = ''     ,
+                   name      = ''     ,
+                   title     = ''     , 
+                   trivial   = False  , 
+                   use_frame = 100000 ) :
+
+        MakeDSTask.__init__ ( self                  ,
+                              variables = variables , 
+                              selection = selection ,
+                              roo_cuts  = roo_cuts  ,
+                              name      = name      ,
+                              title     = title     )
+        
+        self.cuts      = cuts     
+        self.trivial   = trivial  
+        self.use_frame = use_frame 
+
+    ## the actual processing 
+    def process ( self , jobid , item ) :
+
+        import ROOT
+        from ostap.logger.logger import logWarning
+        with logWarning() :
+            import ostap.core.pyrouts            
+            import ostap.trees.trees
+            import ostap.fitting.roofit 
         
         ## reconstruct chain from the item 
         chain    = item.chain
@@ -73,65 +216,30 @@ class  FillTask(Task) :
 
         all = 0 == first and ( nevents < 0 or ll <= nevents )
         
-        if self.trivial and all : 
+        if self.trivial and all and not self.cuts : 
             import ostap.fitting.pyselectors
-            self.__output = chain.make_dataset ( self.variables , self.selection , silent = True ) 
-            return self.__output 
-
+            return chain.make_dataset ( self.variables , self.selection , silent = True ) 
+        
         from   ostap.fitting.pyselectors import SelectorWithVars
         
         ## use selector  
         selector = SelectorWithVars ( variables = self.variables ,
                                       selection = self.selection ,
-                                      roo_cuts  = self.roo_cuts  , 
+                                      roo_cuts  = self.roo_cuts  ,
+                                      cuts      = self.cuts      , 
                                       silence   = True           )
         
         args = ()  
         if not all : args  = nevents , first 
         
-        dataset , stat  = chain.fill_dataset2 ( selector  ,
-                                                *args     ,
-                                                silent    = True                 , 
-                                                shortcut  = all and self.trivial ,
-                                                use_frame = self.use_frame       )
-        
-        ## self.__output = selector.data, selector.stat  
-        self.__output = dataset, stat 
-        
-        ## if  num < 0 :
-        ##   logger.warning ("Processing status %s (jobid #%s)"  %  ( num % jobid ) ) 
-        
-        ## del selector.data
-        ## del      selector        
-        logger.debug ( 'Processed %s and filled %d entries (jobid #%s)' % ( stat , len ( dataset ) , jobid ) )
+        return chain.fill_dataset2 ( selector  ,
+                                     *args     ,
+                                     silent    = True                 , 
+                                     shortcut  = all and self.trivial ,
+                                     use_frame = self.use_frame       )
 
-        return self.__output 
 
-    ## merge results/datasets 
-    def merge_results ( self , result , jobid = -1 ) :
-        
-        import ostap.fitting.dataset
-        if result :
-            ds , stat = result
-            if not self.__output or not self.__output[0] :
-                self.__output = ds , stat  
-            else :
-                ds_ , stat_ = self.__output
-                ds_.append ( ds )
-                stat_.total      += stat.total     ## total 
-                stat_.processed  += stat.processed ## procesed 
-                stat_.skipped    += stat.skipped   ## skipped                
-                self.__output = ds_ , stat_
-                ds.erase () 
-            del result            
-            logger.debug ( 'Merging: %d entries ' % len( self.__output[0] ) )
-        else :
-            logger.error ( "No valid results for merging" )
-            
-    ## get the results 
-    def results ( self ) :
-        return self.__output
-    
+
 # ===================================================================================
 ## parallel processing of loooong chain/tree 
 #  @code
@@ -153,7 +261,7 @@ def parallel_fill ( chain                  ,
     """ Parallel processing of loooong chain/tree 
     >>>chain    = ...
     >>> selector =  ...
-    >>> chain.pprocess ( selector )
+    >>> chain.parallel_fill ( selector )
     """
     import ostap.fitting.roofit 
     from   ostap.fitting.pyselectors import SelectorWithVars 
@@ -172,21 +280,22 @@ def parallel_fill ( chain                  ,
     
     trivial   = selector.really_trivial and not selector.morecuts 
     
-    all = 0 == first and ( 0 > nevents or len ( chain ) <= nevents )
+    all = ( 0 == first ) and ( 0 > nevents or len ( chain ) <= nevents )
     
-    if all and trivial and 1 < len( ch.files ) :
-        logger.info ("Configuration is ``trivial'': redefine ``chunk-size'' to -1")
+    if all and trivial and 1 < len ( ch.files ) :
+        logger.info ("Configuration is `trivial': redefine `chunk-size' to -1")
         chunk_size = -1
         
-    task  = FillTask ( variables = variables ,
-                       selection = selection ,
-                       roo_cuts  = roo_cuts  ,
-                       trivial   = trivial   ,
-                       use_frame = use_frame )
+    task  = FillTask ( variables = variables         ,
+                       selection = selection         ,
+                       cuts      = selector.morecuts , 
+                       roo_cuts  = roo_cuts          ,
+                       trivial   = trivial           ,
+                       use_frame = use_frame         ) 
     
     wmgr  = WorkManager ( silent     = silent     , **kwargs )
     trees = ch.split    ( chunk_size = chunk_size , max_files = max_files )
-    wmgr.process( task , trees , chunk_size = job_chunk )
+    wmgr.process ( task , trees , chunk_size = job_chunk )
     del trees
     
     dataset, stat = task.results()  
@@ -209,6 +318,148 @@ def parallel_fill ( chain                  ,
 
 ROOT.TChain.parallel_fill = parallel_fill 
 ROOT.TTree .parallel_fill = parallel_fill 
+
+
+# =============================================================================
+## Create RooDataset from the chain/tree
+#  @code 
+#  tree = ...
+#  ds, stat = tree.pfill_dataset ( [ 'px , 'py' , 'pz' ] ) 
+#  @endcode
+def parallel_fill_dataset  ( chain                  ,
+                             variables              ,  ## list of variables 
+                             selection    = ''      ,  ## TTree-cuts 
+                             roo_cuts     = ''      ,  ## RooFit cuts 
+                             cuts         = None    ,  ## python callable 
+                             name         = ''      ,
+                             title        = ''      ,
+                             shortcut     = True    ,
+                             use_frame    = 20000   ,   ## important 
+                             silent       = True    , 
+                             max_files    = 1       ,
+                             job_chunk    = -1      , **kwargs ) :
+    """ Create RooDataset from the chain/tree
+    >>> tree = ...
+    >>> ds , stat = tree.pfill_dataset ( [ 'px , 'py' , 'pz' ] )
+    """
+    
+    import ostap.fitting.roofit 
+    from   ostap.fitting.pyselectors import SelectorWithVars 
+    from   ostap.trees.trees         import Chain
+    
+    ch = Chain ( chain ) 
+    if ch.nFiles < 2 :
+        from ostap.fitting.pyselectors import fill_dataset1 as _fill_dataset_
+        return _fill_dataset_ ( tree                  ,
+                                variables = variables , 
+                                selection = selection ,                                
+                                roo_cut   = roo_cuts  ,
+                                cuts      = cuts      ,
+                                name      = name      ,
+                                title     = title     ,
+                                shortcut  = shortcut  ,
+                                use_frame = use_frame ,
+                                silent    = silent    )
+    
+    task  = FillDSTask ( variables = variables  ,
+                         selection = selection  ,
+                         cuts      = cuts       , 
+                         roo_cuts  = roo_cuts   ,
+                         name      = name       ,
+                         title     = title      ,
+                         shortcut  = shortcut   ,
+                         use_frame = use_frame  )
+
+    
+    wmgr  = WorkManager ( silent     = silent     , **kwargs )
+    trees = ch.split    ( chunk_size = -1 , max_files = max_files )
+    wmgr.process ( task , trees , chunk_size = job_chunk )
+    del trees
+    
+    dataset, stat = task.results()  
+
+    from ostap.logger.logger import attention 
+    skipped = 'Skipped:%d' % stat.skipped
+    skipped = '/' + attention ( skipped ) if stat.skipped else ''
+    logger.info (
+        'Selector(%s): Events Processed:%d/Total:%d%s CUTS: "%s"\n%s' % (
+        selector.name    ,
+        stat.processed   ,
+        stat.total       ,
+        skipped          ,
+        selector.cuts()  , dataset.table ( prefix = '# ' ) ) )             
+    
+    return dataset, stat  
+
+
+ROOT.TChain.pfill_dataset = parallel_fill_dataset 
+ROOT.TTree .pfill_dataset = parallel_fill_dataset
+
+# =============================================================================
+## Create RooDataset from the tree using parallel Tree->Frame->Dataset transformation
+#  @code 
+#  tree = ...
+#  ds   = tree.pmake_dataset ( [ 'px , 'py' , 'pz' ] ) 
+#  @endcode
+def parallel_make_dataset ( chain                ,
+                            variables            , ## variables 
+                            selection  = ''      , ## TTree selection 
+                            roo_cuts   = ''      , ## Roo-Fit selection  
+                            name       = ''      , 
+                            title      = ''      ,
+                            silent     =  False  , 
+                            max_files  =  1      ,
+                            job_chunk  = -1      , **kwargs ) :
+    """ Create RooDataset from the tree using parallel Tree->Frame->Dataset transformation 
+    >>> tree = ...
+    >>> ds   = tree.pmake_dataset ( [ 'px , 'py' , 'pz' ] ) 
+    """
+    
+    import ostap.fitting.roofit 
+    from   ostap.trees.trees         import Chain
+    
+    ch = Chain ( chain )
+    
+    if ch.nFiles < 2  :
+        from ostap.fitting.pyselectors import make_dataset as _make_dataset_
+        return _make_dataset_ ( tree                  ,
+                                variables = variables , 
+                                selection = selection ,
+                                roo_cut   = roo_cuts  ,
+                                name      = name      ,
+                                title     = title     ,
+                                silent    = silent    ) 
+                                
+    # and here we have parallel processing 
+
+    task = MakeDSTask ( variables = variables ,
+                        selection = selection , 
+                        roo_cuts  = roo_cuts  ,
+                        name      = name      ,
+                        title     = title     )
+    
+    
+    wmgr  = WorkManager ( silent    = silent , **kwargs )
+    trees = ch.split    ( chunk_size = -1 , max_files = max_files )
+    wmgr.process ( task , trees , chunk_size = job_chunk )
+    del trees
+
+    dataset, stat = task.results()  
+
+    from ostap.logger.logger import attention 
+    skipped = 'Skipped:%d' % stat.skipped
+    skipped = '/' + attention ( skipped ) if stat.skipped else ''
+    logger.info (
+        'Events Processed:%d/Total:%d%s CUTS: "%s"\n%s' % (
+        stat.processed   ,
+        stat.total       ,
+        skipped          ,
+        selection        , dataset.table ( prefix = '# ' ) ) )             
+    
+    return dataset, stat  
+
+ROOT.TChain.pmake_dataset = parallel_make_dataset 
+ROOT.TTree .pmake_dataset = parallel_make_dataset
 
 
 # ===================================================================================
@@ -250,8 +501,12 @@ _decorated_classes_ = (
 _new_methods_       = (
     ROOT.TTree .parallel_fill ,
     ROOT.TChain.parallel_fill ,
-    ROOT.TTree .pprocess ,
-    ROOT.TChain.pprocess ,
+    ROOT.TTree .pfill_dataset ,
+    ROOT.TChain.pfill_dataset ,
+    ROOT.TTree .pmake_dataset ,
+    ROOT.TChain.pmake_dataset ,
+    ROOT.TTree .pprocess      ,
+    ROOT.TChain.pprocess      ,
     )
 
 # =============================================================================